서포터 스킬 구현
서포터 스킬 구현
main scene에서 서포터와 탱커 파라미터 수정
탱커 파라미터의 데이터타입 수정

+++ Assets/Scripts/Tanker.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class Tanker : Ally
 {
     public int shieldAmount;
-    public int shieldTime;
+    public float shieldTime;
     protected override IEnumerator AutoAttack()
     {
         attackRemainTime = delay;
